break up dimension test suite
