+++ tests/dimension_test.rs
@@ -199,72 +199,6 @@ fn it_shows_dimensions_are_quarter_turns() {
 }
 
 #[test]
-fn it_proves_grade_decomposition_ignores_angle_addition() {
-    // traditional geometric algebra ignores that multiplication adds angles
-    // when you multiply v1 * v2, the angles add: θ1 + θ2
-    // but traditional GA pretends this angle addition doesnt happen
-
-    // example: multiply two 45° vectors
-    // v1 at 45°, v2 at 45°
-    // v1 * v2 rotates by 45° + 45° = 90°
-
-    // but traditional GA ignores this simple angle addition and instead:
-    // 1. computes a "scalar part" (grade 0)
-    // 2. computes a "bivector part" (grade 2)
-    // 3. stores both in separate memory locations
-    // 4. pretends the 90° rotation is somehow split between them
-
-    // this negligence - ignoring angle addition - forces traditional GA to:
-    // - track 2^n components to handle all possible angle accumulations
-    // - invent "grade decomposition" to duplicate the angle information
-    // - create massive computational overhead for simple rotations
-
-    // geonum acknowledges that multiplication adds angles
-    let v1 = Geonum::new(1.0, 1.0, 4.0); // 45° = π/4
-    let v2 = Geonum::new(1.0, 1.0, 4.0); // 45° = π/4
-
-    let product = v1 * v2;
-
-    // result: 45° + 45° = 90° rotation, stored as single angle
-    assert_eq!(product.mag, 1.0);
-    assert_eq!(product.angle.blade(), 1); // 90° rotation (blade 1)
-    assert!(product.angle.near_rem(0.0)); // exactly π/2
-
-    // geonum stores the angle addition result directly
-    // no need to decompose into "scalar" and "bivector" parts
-    // no need for 2^n components to track angle accumulations
-
-    // traditional GA creates "grade 0" and "grade 2" components because
-    // it refuses to acknowledge that angles simply added to 90°
-
-    // demonstration: multiply 0° by 90°
-    let x_axis = Geonum::create_dimension(1.0, 0); // 0°
-    let y_axis = Geonum::create_dimension(1.0, 1); // 90°
-
-    let xy_product = x_axis * y_axis;
-
-    // angle addition: 0° + 90° = 90°
-    assert_eq!(xy_product.angle.blade(), 1); // 90° rotation
-    assert!(xy_product.angle.near_rem(0.0));
-
-    // traditional GA would ignore this angle addition and instead:
-    // - compute x·y = 0 (call it "scalar part")
-    // - compute x∧y = 1 (call it "bivector part")
-    // - store both separately
-    // - pretend the 90° rotation is somehow "decomposed"
-
-    // but the 90° rotation hasnt been decomposed - its been ignored!
-    // grade decomposition is what you get when you refuse to track angle addition
-
-    // by ignoring "angles add", traditional GA creates exponential complexity
-    // every possible angle sum needs its own storage location
-    // thats why you get 2^n components - one for each possible accumulation
-
-    // geonum eliminates slack from the geometry by requiring angle addition
-    // no duplication, no exponential blowup, just store the angle sum directly
-}
-
-#[test]
 fn it_proves_vectors_can_never_be_orthogonal() {
     // traditional math inherited the notion that vectors can be "orthogonal"
     // this language emerged from "grade decomposition"
@@ -338,646 +272,6 @@ fn it_proves_vectors_can_never_be_orthogonal() {
     // geonum reveals the underlying pattern: PI/2 rotation changes grade
     // "orthogonal vectors" emerged from overlooking grade transformation
     // the angle-blade invariant makes explicit what grade decomposition obscured
-}
-
-#[test]
-fn it_demonstrates_inversion_preserves_grade_parity_relationships() {
-    // geonum's grade structure has involutive pairs: 0↔2, 1↔3
-    // operations that preserve this pairing maintain orthogonality relationships
-    // circular inversion is one such operation
-    let center = Geonum::new_from_cartesian(0.0, 0.0); // origin for clarity
-    let radius = 2.0;
-
-    // test points at different angles and distances
-    let test_configs = vec![
-        // (distance, angle_pi_rad, angle_div, description)
-        (1.0, 0.0, 1.0, "inside on +x axis"),
-        (3.0, 0.0, 1.0, "outside on +x axis"),
-        (1.0, 1.0, 2.0, "inside on +y axis"),
-        (3.0, 1.0, 2.0, "outside on +y axis"),
-        (1.0, 1.0, 4.0, "inside at π/4"),
-        (3.0, 1.0, 4.0, "outside at π/4"),
-        (1.0, 1.0, 1.0, "inside on -x axis"),
-        (3.0, 1.0, 1.0, "outside on -x axis"),
-    ];
-
-    println!("\nSingle point inversions from origin:");
-    for (dist, pi_rad, div, desc) in test_configs {
-        let p = Geonum::new(dist, pi_rad, div);
-        let p_inv = p.invert_circle(&center, radius);
-
-        println!(
-            "{}: dist={} angle={:.3} blade={} → dist={:.3} angle={:.3} blade={}",
-            desc,
-            dist,
-            p.angle.rem(),
-            p.angle.blade(),
-            p_inv.mag,
-            p_inv.angle.rem(),
-            p_inv.angle.blade()
-        );
-
-        // verify inversion property
-        assert!((p.mag * p_inv.mag - radius * radius).abs() < EPSILON);
-    }
-
-    // now test difference vectors between points (where blade changes occurred before)
-    println!("\nDifference vectors between points:");
-
-    // create a configuration that shows blade transformation
-    let p1 = Geonum::new_from_cartesian(2.0, 1.0);
-    let p2 = Geonum::new_from_cartesian(3.0, 0.0);
-    let p3 = Geonum::new_from_cartesian(2.0, -1.0);
-
-    // compute difference vectors
-    let v12 = p2 - p1;
-    let v13 = p3 - p1;
-    let v23 = p3 - p2;
-
-    println!("Original vectors:");
-    println!(
-        "  v12=p2-p1: length={:.3} angle={:.3} blade={}",
-        v12.mag,
-        v12.angle.rem(),
-        v12.angle.blade()
-    );
-    println!(
-        "  v13=p3-p1: length={:.3} angle={:.3} blade={}",
-        v13.mag,
-        v13.angle.rem(),
-        v13.angle.blade()
-    );
-    println!(
-        "  v23=p3-p2: length={:.3} angle={:.3} blade={}",
-        v23.mag,
-        v23.angle.rem(),
-        v23.angle.blade()
-    );
-
-    // invert the points
-    let p1_inv = p1.invert_circle(&center, radius);
-    let p2_inv = p2.invert_circle(&center, radius);
-    let p3_inv = p3.invert_circle(&center, radius);
-
-    // compute inverted difference vectors
-    let v12_inv = p2_inv - p1_inv;
-    let v13_inv = p3_inv - p1_inv;
-    let v23_inv = p3_inv - p2_inv;
-
-    println!("Inverted vectors:");
-    println!(
-        "  v12_inv: length={:.3} angle={:.3} blade={}",
-        v12_inv.mag,
-        v12_inv.angle.rem(),
-        v12_inv.angle.blade()
-    );
-    println!(
-        "  v13_inv: length={:.3} angle={:.3} blade={}",
-        v13_inv.mag,
-        v13_inv.angle.rem(),
-        v13_inv.angle.blade()
-    );
-    println!(
-        "  v23_inv: length={:.3} angle={:.3} blade={}",
-        v23_inv.mag,
-        v23_inv.angle.rem(),
-        v23_inv.angle.blade()
-    );
-
-    // KEY INSIGHT: blade transformation happens in difference vectors
-    // individual points from origin maintain blade, but vectors between inverted points transform
-
-    // test with points that create perpendicular vectors
-    println!("\nPerpendicular vector configuration:");
-    let center2 = Geonum::new_from_cartesian(1.0, 0.0); // offset center
-    let q1 = Geonum::new_from_cartesian(3.0, 0.0);
-    let q2 = Geonum::new_from_cartesian(4.0, 0.0);
-    let q3 = Geonum::new_from_cartesian(3.0, 1.0);
-
-    let u1 = q2 - q1; // horizontal
-    let u2 = q3 - q1; // vertical
-
-    println!("Original perpendicular vectors:");
-    println!("  u1: blade={} (horizontal)", u1.angle.blade());
-    println!("  u2: blade={} (vertical)", u2.angle.blade());
-
-    // these perpendicular vectors have different blades (orthogonality via blade difference)
-    assert_ne!(
-        u1.angle.blade() % 2,
-        u2.angle.blade() % 2,
-        "perpendicular vectors differ by odd blade count"
-    );
-
-    let q1_inv = q1.invert_circle(&center2, radius);
-    let q2_inv = q2.invert_circle(&center2, radius);
-    let q3_inv = q3.invert_circle(&center2, radius);
-
-    let u1_inv = q2_inv - q1_inv;
-    let u2_inv = q3_inv - q1_inv;
-
-    println!("Inverted 'perpendicular' vectors:");
-    println!("  u1_inv: blade={}", u1_inv.angle.blade());
-    println!("  u2_inv: blade={}", u2_inv.angle.blade());
-
-    // blade relationships transform under inversion
-    let blade_diff_original = (u2.angle.blade() as i32 - u1.angle.blade() as i32).abs();
-    let blade_diff_inverted = (u2_inv.angle.blade() as i32 - u1_inv.angle.blade() as i32).abs();
-
-    println!("Blade difference: {blade_diff_original} → {blade_diff_inverted}");
-
-    // check if grade differences are preserved (blade mod 4)
-    let grade_diff_original =
-        ((u2.angle.grade() as i32 - u1.angle.grade() as i32).abs() % 4) as usize;
-    let grade_diff_inverted =
-        ((u2_inv.angle.grade() as i32 - u1_inv.angle.grade() as i32).abs() % 4) as usize;
-
-    println!("Grade difference: {grade_diff_original} → {grade_diff_inverted}");
-
-    // orthogonality is encoded in odd grade differences (parity)
-    // grade 0 vs grade 1: difference = 1 (odd) → orthogonal
-    // grade 2 vs grade 3: difference = 1 (odd) → orthogonal
-    // grade 0 vs grade 2: difference = 2 (even) → parallel (dual pair)
-    // grade 1 vs grade 3: difference = 2 (even) → parallel (dual pair)
-
-    assert_eq!(
-        grade_diff_original % 2,
-        1,
-        "original vectors are orthogonal (odd grade diff)"
-    );
-    assert_eq!(
-        grade_diff_inverted % 2,
-        1,
-        "inverted vectors remain orthogonal (odd grade diff)"
-    );
-
-    // this is expected from geonum's involutive grade pairs (0↔2, 1↔3)
-    // operations respecting this pairing preserve orthogonality parity
-    // circular inversion is such an operation - it may shift grades within pairs
-    // but preserves the odd/even nature of grade differences
-}
-
-#[test]
-fn it_solves_the_exponential_complexity_explosion() {
-    // THE PROBLEM: traditional GA suffers from 2^n explosion
-    // why? it refuses to acknowledge that rotations compose by angle addition
-    // instead, it scatters rotation information across exponentially many components
-
-    // traditional GA component count:
-    // 1D: 2 components (scalar, e1)
-    // 2D: 4 components (scalar, e1, e2, e12)
-    // 3D: 8 components (scalar, e1, e2, e3, e12, e13, e23, e123)
-    // 10D: 1024 components (all possible products of basis vectors)
-    // nD: 2^n components
-
-    // THE SOLUTION: geonum recognizes that a rotation is just [length, angle]
-    // no matter how many dimensions, a 45° rotation is stored as one number
-
-    // proof: represent a 45° rotation
-    let rotation_45 = Geonum::new(1.0, 1.0, 4.0); // [1, π/4]
-
-    // apply this rotation to different objects - always the same operation
-    let x_axis = Geonum::new(1.0, 0.0, 1.0);
-    let rotated = x_axis * rotation_45; // rotate x-axis by 45°
-    assert_eq!(rotated.angle, Angle::new(1.0, 4.0)); // now at 45°
-
-    // this single number works in ANY dimension:
-    // - in 2D: rotates in the xy-plane
-    // - in 3D: rotates in the xy-plane (z unchanged)
-    // - in 10D: rotates in the xy-plane (other 8 dims unchanged)
-
-    // traditional GA cant do this! it needs:
-    // - 2D: distribute across 4 components
-    // - 3D: distribute across 8 components
-    // - 10D: distribute across 1024 components
-    // all to represent the same simple 45° rotation
-
-    // the key insight: multiplication is just angle addition
-    let a = Geonum::new(2.0, 1.0, 6.0); // [2, π/6]
-    let b = Geonum::new(3.0, 1.0, 3.0); // [3, π/3]
-    let product = a * b;
-
-    // geonum: O(1) operations
-    assert_eq!(product.mag, 6.0); // lengths multiply
-    assert_eq!(product.angle, Angle::new(1.0, 2.0)); // angles add: π/6 + π/3 = π/2
-
-    // traditional GA: O(4^n) operations for the same result!
-    // in 10D: 1024 × 1024 = 1,048,576 component multiplications
-    // of which 99.9% produce zeros that still get computed and stored
-
-    // even worse: chained operations
-    let c = Geonum::new(1.5, 1.0, 4.0); // [1.5, π/4]
-    let chain = a * b * c;
-
-    // geonum: still O(1)
-    assert_eq!(chain.mag, 9.0); // 2 × 3 × 1.5
-    assert_eq!(chain.angle, Angle::new(3.0, 4.0)); // π/6 + π/3 + π/4 = 3π/4
-
-    // traditional GA: must expand (a*b) into 2^n components,
-    // then multiply all 2^n by c's 2^n components
-    // the explosion compounds with every operation!
-
-    // geonum solves the 2^n explosion by storing what actually matters:
-    // the total rotation angle, not its decomposition into 2^n pieces
-}
-
-/// test proving that geonum eliminates the need for pseudoscalars
-///
-/// Traditional geometric algebra requires pseudoscalars (like e₁∧e₂∧e₃ in 3D) to:
-/// 1. Define duality operations: dual(A) = A * I where I is the pseudoscalar
-/// 2. Represent oriented volume elements
-/// 3. Handle metric and orientation of the space
-/// 4. Define the "top grade" of multivectors
-///
-/// Geonum's angle-blade representation makes pseudoscalars unnecessary
-/// by encoding these relationships directly in the geometric structure.
-
-#[test]
-fn it_doesnt_need_a_pseudoscalar() {
-    // 1. DUALITY MAPPING: v* = v · I (multiply by pseudoscalar)
-    //    traditional GA needs to:
-    //    - define basis vectors e₁, e₂, e₃
-    //    - compute pseudoscalar I = e₁∧e₂∧e₃
-    //    - multiply vector by I to get dual
-    //    geonum skips all that - dual() just adds 2 blades directly
-    let vector = Geonum::new_with_blade(1.0, 1, 0.0, 1.0); // blade 1 = vector
-    let dual = vector.dual();
-    assert_eq!(dual.angle.blade(), 3); // blade 1 + 2 = 3 (trivector)
-
-    // 2. VOLUME ORIENTATION: I² = ±1 determines metric signature
-    //    traditional: must compute pseudoscalar square to determine orientation
-    //    geonum: dual operation handles orientation through blade arithmetic
-    let volume = Geonum::new_with_blade(1.0, 3, 0.0, 1.0); // grade 3 volume element
-    let dual_volume = volume.dual(); // duality transformation without pseudoscalar
-    assert_eq!(dual_volume.angle.blade(), 5); // blade 3 + 2 = 5
-    assert_eq!(dual_volume.angle.grade(), 1); // grade 5 % 4 = 1 (vector)
-                                              // proves duality maps grade 3→1 without constructing I or computing I²
-
-    // 3. HODGE STAR OPERATION: *v = v · I / |I|²
-    //    traditional: matrix transformation + magnitude computation and verification
-    //    geonum: direct blade transformation
-    let vector = [3.0, 0.0, 0.0];
-    let pseudoscalar_transform = [
-        [0.0, 0.0, 1.0],  // x → yz dual
-        [0.0, 0.0, -1.0], // y → zx dual
-        [1.0, 0.0, 0.0],  // z → xy dual
-    ];
-    let mut dual_vector = [0.0; 3];
-    for i in 0..3 {
-        for (j, &vector_j) in vector.iter().enumerate() {
-            dual_vector[i] += pseudoscalar_transform[i][j] * vector_j;
-        }
-    }
-    // traditional hodge star requires magnitude verification: |*v| = |v|
-    let vector_magnitude_sq: f64 =
-        vector[0] * vector[0] + vector[1] * vector[1] + vector[2] * vector[2];
-    let dual_magnitude_sq: f64 = dual_vector[0] * dual_vector[0]
-        + dual_vector[1] * dual_vector[1]
-        + dual_vector[2] * dual_vector[2];
-    assert_eq!(vector_magnitude_sq.sqrt(), dual_magnitude_sq.sqrt()); // |I|² normalization preserved
-    assert_eq!(dual_vector, [0.0, 0.0, 3.0]);
-
-    // 4. GRADE EXTRACTION: extract k-grade part using pseudoscalar projections
-    //    traditional: pseudoscalar projections for grade filtering
-    //    geonum: unified operations eliminates grade extraction dependency
-    let scale_factor = 2.0; // scalar component (grade 0)
-    let rotation_angle = PI / 4.0; // bivector component (grade 2 via rotation)
-
-    // traditional: extract grades using pseudoscalar projections
-    // grade extraction formula: ⟨A⟩ₖ = (A ∧ Iᵏ) ⌋ I⁻ᵏ / k!
-
-    // extract scalar (grade 0): ⟨A⟩₀ = A ⌋ I⁰ = A ⌋ 1 = scalar_part
-    let scalar_part = scale_factor; // grade 0 extraction
-    let pseudoscalar_grade0 = 1.0; // I⁰ = identity for scalar extraction
-    assert_eq!(scalar_part * pseudoscalar_grade0, scale_factor); // pseudoscalar I⁰ appears in extraction
-
-    // extract bivector (grade 2): ⟨A⟩₂ = (A ∧ I²) ⌋ I⁻² / 2!
-    // in 2D: pseudoscalar I = e₁∧e₂, so I² = -1
-    let pseudoscalar_2d = -1.0; // I² = (e₁∧e₂)² = -1 in 2D
-    let bivector_coefficient = rotation_angle / pseudoscalar_2d; // divide by I² for extraction
-    let rotation_matrix = [
-        [rotation_angle.cos(), -rotation_angle.sin()],
-        [rotation_angle.sin(), rotation_angle.cos()],
-    ]; // grade 2 (bivector) as rotation matrix
-    assert_eq!(bivector_coefficient, -rotation_angle); // pseudoscalar I² = -1 negates coefficient in extraction formula
-
-    // apply decomposed operations: scaling then rotation matrix multiplication
-    let input_point = [1.0, 0.0];
-    let scaled_point = [input_point[0] * scalar_part, input_point[1] * scalar_part];
-    let traditional_result = [
-        rotation_matrix[0][0] * scaled_point[0] + rotation_matrix[0][1] * scaled_point[1],
-        rotation_matrix[1][0] * scaled_point[0] + rotation_matrix[1][1] * scaled_point[1],
-    ];
-
-    // geonum: unified operation without grade extraction
-    let input = Geonum::new(1.0, 0.0, 1.0);
-    let geonum_result = input.scale_rotate(scale_factor, Angle::new(rotation_angle, PI));
-    let geonum_x = geonum_result.mag * geonum_result.angle.grade_angle().cos();
-    let geonum_y = geonum_result.mag * geonum_result.angle.grade_angle().sin();
-    assert!((traditional_result[0] - geonum_x).abs() < 1e-10);
-    assert!((traditional_result[1] - geonum_y).abs() < 1e-10);
-
-    // 5. COMPLEMENT OPERATIONS: orthogonal complement via pseudoscalar
-    //    traditional: multiply by pseudoscalar for complement A^⊥ = A · I
-    //    geonum: complement through blade arithmetic (dual operation)
-
-    // traditional: construct pseudoscalar I for complement operation
-    let pseudoscalar_3d = 1.0; // I₃ = e₁∧e₂∧e₃ = +1 in right-handed 3D
-    let line_vector = [1.0, 0.0, 0.0]; // line in x direction
-
-    // traditional complement: A^⊥ = A · I (multiply by pseudoscalar)
-    let traditional_complement = [
-        line_vector[0] * pseudoscalar_3d,
-        line_vector[1] * pseudoscalar_3d,
-        line_vector[2] * pseudoscalar_3d,
-    ]; // complement is the orthogonal subspace
-    assert_eq!(traditional_complement, [1.0, 0.0, 0.0]);
-
-    // geonum: complement through dual() blade arithmetic
-    let line = Geonum::new_with_blade(1.0, 1, 0.0, 1.0); // grade 1 line
-    let complement = line.dual(); // complement via blade addition
-    assert_eq!(complement.angle.blade(), 3); // grade 1 → grade 3 (line → volume)
-    assert_eq!(complement.mag, 1.0); // magnitude preserved
-
-    // 6. CROSS PRODUCTS: v × w = (v ∧ w) · I^(-1) in 3D
-    //    traditional: wedge then multiply by pseudoscalar inverse
-    //    geonum: wedge product directly
-    let v = [1.0, 0.0, 0.0]; // x unit vector
-    let w = [0.0, 1.0, 0.0]; // y unit vector
-
-    // traditional: compute wedge product v∧w
-    let e12_coeff = v[0] * w[1] - v[1] * w[0]; // xy-component = 1
-    let e23_coeff = v[1] * w[2] - v[2] * w[1]; // yz-component = 0
-    let e31_coeff = v[2] * w[0] - v[0] * w[2]; // zx-component = 0
-
-    // traditional: left-handed pseudoscalar I₃ = -1 (solution to sign error)
-    let pseudoscalar_i3 = -1.0; // left-handed orientation
-    let i3_squared = -1.0; // (e₁∧e₂∧e₃)² = -1 in 3D euclidean GA
-    let pseudoscalar_inverse = pseudoscalar_i3 / i3_squared; // (-1) / (-1) = +1
-
-    // traditional: cross product (v∧w) · I₃⁻¹
-    let traditional_cross = [
-        e23_coeff * pseudoscalar_inverse, // yz → x component
-        e31_coeff * pseudoscalar_inverse, // zx → y component
-        e12_coeff * pseudoscalar_inverse, // xy → z component
-    ];
-    assert_eq!(traditional_cross, [0.0, 0.0, 1.0]); // x × y = z
-
-    // test magnitude |v × w| = |v| |w| sin(θ)
-    let cross_magnitude_sq: f64 = traditional_cross[0] * traditional_cross[0]
-        + traditional_cross[1] * traditional_cross[1]
-        + traditional_cross[2] * traditional_cross[2];
-    let cross_magnitude = cross_magnitude_sq.sqrt();
-    let v_magnitude_sq: f64 = v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
-    let v_magnitude = v_magnitude_sq.sqrt();
-    let w_magnitude_sq: f64 = w[0] * w[0] + w[1] * w[1] + w[2] * w[2];
-    let w_magnitude = w_magnitude_sq.sqrt();
-    let expected_magnitude = v_magnitude * w_magnitude * (PI / 2.0).sin(); // sin(90°) = 1
-    assert_eq!(cross_magnitude, expected_magnitude);
-
-    // geonum: wedge without pseudoscalar inverse computation
-    let v_geonum = Geonum::new_from_cartesian(1.0, 0.0);
-    let w_geonum = Geonum::new_from_cartesian(0.0, 1.0);
-    let geonum_wedge = v_geonum.wedge(&w_geonum);
-    assert_eq!(geonum_wedge.mag, 1.0); // magnitude preserved without pseudoscalar operations
-
-    // 7. NORMAL VECTORS: surface normals via pseudoscalar multiplication
-    //    traditional: compute tangent cross product then multiply pseudoscalar
-    //    geonum: normals through grade parity relationships
-
-    // traditional: surface normal = (v1 × v2) · I₃⁻¹ requires pseudoscalar
-    // geonum: surface normal via grade parity - odd differences are orthogonal
-
-    // create surface as bivector (grade 2)
-    let surface_x = Geonum::new(1.0, 0.0, 1.0); // grade 0
-    let surface_y = surface_x.rotate(Angle::new(1.0, 2.0)); // grade 1
-    let surface_plane = surface_x.wedge(&surface_y); // grade 2
-
-    // find normals: grades with odd difference from surface grade 2
-    let normal_vector = surface_y; // grade 1 (diff=1, odd=orthogonal)
-    let normal_trivector = surface_y
-        .rotate(Angle::new(1.0, 2.0))
-        .rotate(Angle::new(1.0, 2.0)); // grade 3 (diff=1, odd=orthogonal)
-    let parallel_scalar = surface_plane.dual(); // grade 0 (diff=2, even=parallel)
-
-    // test orthogonality via dot products
-    let dot_vector = surface_plane.dot(&normal_vector);
-    let dot_trivector = surface_plane.dot(&normal_trivector);
-    let dot_parallel = surface_plane.dot(&parallel_scalar);
-
-    assert!(
-        dot_vector.mag.abs() < 1e-10,
-        "grade 1 normal orthogonal to grade 2 surface"
-    );
-    assert!(
-        dot_trivector.mag.abs() < 1e-10,
-        "grade 3 normal orthogonal to grade 2 surface"
-    );
-    assert!(
-        dot_parallel.mag.abs() > 1e-10,
-        "grade 0 parallel to grade 2 surface (even diff)"
-    );
-
-    // geonum ghosts pseudoscalar I₃ multiplication
-    // orthogonality emerges from grade parity, no pseudoscalar needed
-
-    // CONCLUSION: All traditional pseudoscalar functionality achieved
-    // through direct angle-blade arithmetic, eliminating the need for:
-    // - Special pseudoscalar objects
-    // - Dimension-specific unit volume elements
-    // - Complex duality multiplication formulas
-    // - Grade-dependent pseudoscalar properties
-    // - Metric signature pseudoscalar complications
-}
-
-#[test]
-fn it_demonstrates_pseudoscalar_elimination_benefits() {
-    // traditional GA rotation: R*v*R† requires 3 geometric products × 64 operations = 192 operations
-    // geonum rotation: angle addition = 1 operation
-    // the 192x difference comes from eliminating the pseudoscalar
-
-    // test rotation equivalence
-    let point = Geonum::new_from_cartesian(3.0, 4.0);
-    let rotation = Angle::new(1.0, 2.0); // π/2
-    let rotated = point.rotate(rotation);
-
-    let x = rotated.mag * rotated.angle.grade_angle().cos();
-    let y = rotated.mag * rotated.angle.grade_angle().sin();
-    assert!((x - (-4.0)).abs() < EPSILON);
-    assert!((y - 3.0).abs() < EPSILON);
-
-    // traditional GA needs pseudoscalar to define rotation planes through basis blade products
-    // this forces dimension-specific formulas and full multiplication tables
-    // geonum recognizes rotation IS angle addition, not basis blade reconstruction
-
-    // test rotation composition
-    let r1 = Angle::new(1.0, 4.0); // π/4
-    let r2 = Angle::new(1.0, 6.0); // π/6
-    let r3 = Angle::new(1.0, 3.0); // π/3
-
-    // traditional: R1*R2*R3 through geometric products
-    // geonum: r1 + r2 + r3
-    let composed = r1 + r2 + r3;
-    assert_eq!(composed, Angle::new(3.0, 4.0)); // 3π/4
-
-    let final_point = point.rotate(composed);
-    let fx = final_point.mag * final_point.angle.grade_angle().cos();
-    let fy = final_point.mag * final_point.angle.grade_angle().sin();
-    assert!((fx - (-4.949)).abs() < 0.01);
-    assert!((fy - (-0.707)).abs() < 0.01);
-
-    // eliminating pseudoscalar reveals rotation as angle addition
-    // 192 operations → 1 operation
-}
-
-#[test]
-fn it_proves_dualization_as_angle_ops_compresses_ga() {
-    // traditional GA duality requires multiplying by dimension-specific pseudoscalars
-    // with exponential 2^n component arrays
-    // geonum reduces duality to O(1) angle arithmetic through the quadrature's bivector
-
-    // the 4 scalar, vector, bivector, trivector principle grades
-    // emerge from the quadrature's bivector: sin(θ+π/2) = cos(θ)
-    // this π/2 rotation creates the 4-quarter-turn cycle needed for complete GA
-
-    // demonstrate duality compression: any dimensional object → simple angle operation
-    let million_dim_vector = Geonum::new_with_blade(1.0, 1_000_000, 1.0, 4.0);
-    let billion_dim_bivector = Geonum::new_with_blade(2.0, 1_000_000_000, 1.0, 3.0);
-
-    // traditional GA: needs 2^1000000 and 2^1000000000 components for duality
-    // (more storage than atoms in observable universe)
-
-    // geonum: duality is just angle arithmetic regardless of dimension
-    let dual_million = million_dim_vector.dual();
-    let dual_billion = billion_dim_bivector.dual();
-
-    // duality maps through 4-cycle: grade k → grade (k+2) % 4
-    // dual() adds 2 blades (π rotation)
-    assert_eq!(million_dim_vector.angle.grade(), 0); // 1000000 % 4 = 0 (scalar)
-    assert_eq!(dual_million.angle.grade(), 2); // (0+2) % 4 = 2 (bivector)
-
-    assert_eq!(billion_dim_bivector.angle.grade(), 0); // 1000000000 % 4 = 0 (scalar)
-    assert_eq!(dual_billion.angle.grade(), 2); // (0+2) % 4 = 2 (bivector)
-
-    // compression achieved: exponential → constant time
-    // traditional: O(2^n) storage and computation
-    // geonum: O(1) angle operations from quadrature's bivector foundation
-
-    // the key insight: the bivector sin(θ+π/2) = cos(θ) IS the duality operator
-    // this π/2 rotation imposes the incidence structure that defines geometric relationships
-    // point-line-plane-volume duality emerges from this single trigonometric identity
-    // eliminating need for dimension-specific pseudoscalars or exponential storage
-
-    // prove duality preserves length (isometry property)
-    assert_eq!(dual_million.mag, million_dim_vector.mag);
-    assert_eq!(dual_billion.mag, billion_dim_bivector.mag);
-
-    // prove duality involution: dual(dual(x)) returns to original grade
-    let double_dual_million = dual_million.dual();
-    let double_dual_billion = dual_billion.dual();
-    assert_eq!(
-        double_dual_million.angle.grade(),
-        million_dim_vector.angle.grade()
-    );
-    assert_eq!(
-        double_dual_billion.angle.grade(),
-        billion_dim_bivector.angle.grade()
-    );
-
-    // prove O(1) complexity: blade arithmetic regardless of dimension size
-    // dual() adds 2 blades regardless of dimension
-    assert_eq!(
-        dual_million.angle.blade(),
-        million_dim_vector.angle.blade() + 2
-    );
-    assert_eq!(
-        dual_billion.angle.blade(),
-        billion_dim_bivector.angle.blade() + 2
-    );
-}
-
-#[test]
-fn it_replaces_k_to_n_minus_k_with_k_to_4_minus_k() {
-    // traditional GA: duality maps grade k to grade (n-k) where n = space dimension
-    // different dimensional spaces need different duality mappings
-    // 3D: k → (3-k), 4D: k → (4-k), 1000D: k → (1000-k)
-
-    // geonum: universal duality k → (4-k) % 4 regardless of dimensional space
-    // works for any dimension through quadrature's bivector foundation
-
-    // demonstrate universal mapping across arbitrary dimensions
-    let obj_3d = Geonum::new_with_blade(1.0, 1, 0.0, 1.0); // grade 1 in "3D context"
-    let obj_1000d = Geonum::new_with_blade(1.0, 1001, 0.0, 1.0); // grade 1 in "1000D context"
-    let obj_million_d = Geonum::new_with_blade(1.0, 1_000_001, 0.0, 1.0); // grade 1 in "million-D context"
-
-    // traditional GA would need different formulas:
-    // 3D: grade 1 → grade (3-1) = 2
-    // 1000D: grade 1 → grade (1000-1) = 999
-    // million-D: grade 1 → grade (1000000-1) = 999999
-
-    // geonum uses same formula k → (4-k) % 4 for all:
-    let dual_3d = obj_3d.dual();
-    let dual_1000d = obj_1000d.dual();
-    let dual_million_d = obj_million_d.dual();
-
-    // all grade 1 objects map to grade 3 regardless of "dimensional context"
-    assert_eq!(obj_3d.angle.grade(), 1);
-    assert_eq!(obj_1000d.angle.grade(), 1);
-    assert_eq!(obj_million_d.angle.grade(), 1);
-
-    assert_eq!(dual_3d.angle.grade(), 3); // (1+2) % 4 = 3
-    assert_eq!(dual_1000d.angle.grade(), 3); // (1+2) % 4 = 3
-    assert_eq!(dual_million_d.angle.grade(), 3); // (1+2) % 4 = 3
-
-    // demonstrate grade 2 → grade 0 universally
-    let bivector_any_dim = Geonum::new_with_blade(2.0, 1002, 0.0, 1.0); // grade 2
-    let dual_bivector = bivector_any_dim.dual();
-
-    assert_eq!(bivector_any_dim.angle.grade(), 2);
-    assert_eq!(dual_bivector.angle.grade(), 0); // (2+2) % 4 = 0
-
-    // compression: eliminates dimension-dependent duality formulas
-    // one universal k → (4-k) % 4 mapping works for any dimensional space
-
-    // geonum eliminates binomial coefficient (n choose k) component explosion
-    // traditional GA: 3D needs (3 choose 1) = 3 vectors, 1000D needs (1000 choose 1) = 1000 vectors
-    // geonum: grade 1 objects use same single [length, angle] representation regardless of dimension
-    // "linearly independent k-vectors" are irrelevant - direction exists naturally through angle preservation
-
-    // geonum eliminates Hodge decomposition: ω = dα + δβ + γ
-    // traditional: separate storage for exact, co-exact, and harmonic components with orthogonal projections
-    // geonum: all decomposition distinctions collapse to angle arithmetic
-    let form_omega = Geonum::new_with_blade(1.0, 5, 1.0, 3.0); // arbitrary differential form
-    let exact_component = form_omega.rotate(Angle::new(1.0, 2.0)); // dα becomes π/2 rotation
-    let coexact_component = form_omega.rotate(Angle::new(3.0, 2.0)); // δβ becomes 3π/2 rotation
-    let harmonic_component = form_omega; // γ is original angle relationship
-
-    // prove no separate storage needed for Hodge decomposition components
-    assert_eq!(
-        std::mem::size_of_val(&form_omega),
-        std::mem::size_of_val(&exact_component)
-    );
-    assert_eq!(
-        std::mem::size_of_val(&exact_component),
-        std::mem::size_of_val(&coexact_component)
-    );
-    assert_eq!(
-        std::mem::size_of_val(&coexact_component),
-        std::mem::size_of_val(&harmonic_component)
-    );
-
-    // prove all grade 1 objects have identical storage regardless of "dimensional context"
-    assert_eq!(
-        std::mem::size_of_val(&obj_3d),
-        std::mem::size_of_val(&obj_1000d)
-    );
-    assert_eq!(
-        std::mem::size_of_val(&obj_1000d),
-        std::mem::size_of_val(&obj_million_d)
-    );
-
-    // traditional GA storage would scale with binomial coefficients:
-    // 3D grade 1: 3 components, 1000D grade 1: 1000 components, million-D grade 1: 1000000 components
-    // geonum storage: constant 2 components (length + angle) for any dimension
 }
 
 #[test]
@@ -1125,44 +419,6 @@ fn it_proves_anticommutativity_is_a_geometric_transformation() {
 
     // this reframes anticommutativity from algebraic property to geometric navigation
     // blade arithmetic isn't representing orientation - it IS orientation
-}
-
-#[test]
-fn it_compresses_traditional_ga_grades_to_two_involutive_pairs() {
-    // geonum's π-rotation dual creates a different incidence structure than traditional GA
-    // instead of computing maximal common subspaces, it computes containing spaces
-
-    let line1 = Geonum::new_with_blade(1.0, 1, 0.0, 1.0); // grade 1
-    let line2 = Geonum::new_with_blade(1.0, 1, 1.0, 4.0); // grade 1, different angle
-    let bivector = Geonum::new_with_blade(1.0, 2, 0.0, 1.0); // grade 2
-    let bivector2 = Geonum::new_with_blade(1.0, 2, 1.0, 4.0); // grade 2, different angle
-
-    // line meet line → grade 1 (vector)
-    // geometric meaning: the intersection point represented as a vector from origin
-    // traditional GA expects grade 0 (scalar point)
-    assert_eq!(line1.meet(&line2).angle.grade(), 1);
-
-    // vector meet bivector → grade 2 (bivector)
-    // geometric meaning: the minimal plane containing both the line and the original plane
-    // traditional GA expects grade 0 (point of intersection)
-    assert_eq!(line1.meet(&bivector).angle.grade(), 2);
-
-    // bivector meet bivector → grade 3 (trivector)
-    // geometric meaning: the 3D volume spanned by the two planes
-    // traditional GA expects grade 1 (line of intersection)
-    assert_eq!(bivector.meet(&bivector2).angle.grade(), 3);
-
-    // this reversal happens because π-rotation dual creates scalar↔bivector
-    // and vector↔trivector pairings rather than traditional complementary pairings
-
-    // KEY INSIGHT: geonum flattens traditional GA's n+1 grade levels (0 through n)
-    // to just 2 involutive pairs that work in any dimension:
-    // - pair 1: grade 0 ↔ grade 2 (scalar ↔ bivector)
-    // - pair 2: grade 1 ↔ grade 3 (vector ↔ trivector)
-    //
-    // grades cycle modulo 4, so grade 1000000 in million-D space is just grade 0
-    // this eliminates dimension-specific k→(n-k) duality formulas
-    // replacing them with universal k→(k+2)%4 that works everywhere
 }
 
 #[test]
@@ -1378,41 +634,6 @@ fn it_sets_angle_forward_geometry_as_primitive() {
     // - blade count IS the transformation history
     // - base_angle() lets you forget history when you dont need it
     // - but the primitive geometry remains: angles only go forward
-}
-
-#[test]
-fn it_handles_mixed_grade_operations_naturally() {
-    // traditional GA: restricts operations to "like grades" or requires complex rules
-    // scalar * scalar = scalar, vector * vector = scalar + bivector, etc.
-    // mixed grade operations need special handling and decomposition
-
-    // geonum: blade arithmetic works for ANY grade combination
-    let scalar = Geonum::new(2.0, 0.0, 1.0); // blade 0 (grade 0)
-    let vector = Geonum::new(3.0, 1.0, 2.0); // blade 1 (grade 1)
-    let bivector = Geonum::new(1.5, 1.0, 1.0); // blade 2 (grade 2)
-    let trivector = Geonum::new(4.0, 3.0, 2.0); // blade 3 (grade 3)
-
-    // mixed grade products: blade counts just add
-    let scalar_vector = scalar * vector; // 0+1=1 (vector)
-    let vector_bivector = vector * bivector; // 1+2=3 (trivector)
-    let bivector_trivector = bivector * trivector; // 2+3=5 (grade 1: 5%4=1)
-    let scalar_trivector = scalar * trivector; // 0+3=3 (trivector)
-
-    // verify blade arithmetic works regardless of starting grades
-    assert_eq!(scalar_vector.angle.blade(), 1);
-    assert_eq!(vector_bivector.angle.blade(), 3);
-    assert_eq!(bivector_trivector.angle.blade(), 5);
-    assert_eq!(scalar_trivector.angle.blade(), 3);
-
-    // verify grades cycle correctly (blade % 4)
-    assert_eq!(scalar_vector.angle.grade(), 1); // blade 1 → grade 1
-    assert_eq!(vector_bivector.angle.grade(), 3); // blade 3 → grade 3
-    assert_eq!(bivector_trivector.angle.grade(), 1); // blade 5 → grade 1
-    assert_eq!(scalar_trivector.angle.grade(), 3); // blade 3 → grade 3
-
-    // traditional GA: each combination needs special rules and storage
-    // geonum: universal blade addition works for all grade combinations
-    // no restrictions, no special cases, no decomposition complexity
 }
 
 #[test]
